requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.5
Pillow>=8.3.0
matplotlib>=3.3.4
tqdm>=4.62.0
opencv-python>=4.5.3
